Minor refactoring.  File input.
